# docs/swagger.yaml
definitions:
...
  common.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  task.AssignTaskRequest:
    properties:
      assigned_to:
        type: string
    required:
    - assigned_to
    type: object
  task.CreateTaskRequest:
    properties:
      assigned_to:
        type: string
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      project_id:
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - project_id
    - title
    type: object
  task.TaskListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/task.TaskResponse'
        type: array
      total:
        type: integer
    type: object
  task.TaskResponse:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  task.UpdateTaskRequest:
    properties:
      assigned_to:
        type: string
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/tasks:
    get:
      consumes:
      - application/json
      description: Get tasks with filtering and pagination (role-based access)
      parameters:
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      - description: Filter by assigned user ID
        in: query
        name: assigned_to
        type: string
      - description: Filter by status (pending, in_progress, completed)
        in: query
        name: status
        type: string
      - description: Filter by priority (low, medium, high)
        in: query
        name: priority
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/task.TaskListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get tasks with filters
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Create a new task (must belong to a project)
      parameters:
      - description: Task creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/task.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Project or user not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new task
      tags:
      - Task
  /api/v1/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing task (creator, manager, or admin only)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete task
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: Get a task by its ID (with permission checks)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/task.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get task by ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update an existing task (with permission checks)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/task.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update task
      tags:
      - Task
  /api/v1/tasks/{id}/assign:
    put:
      consumes:
      - application/json
      description: Assign a task to a user (managers and admins only)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task assignment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.AssignTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task assigned successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/task.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Task or user not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Assign task to user
      tags:
      - Task
...
swagger: "2.0"